% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAllTracks.R
\name{combineAllTracks}
\alias{combineAllTracks}
\title{Combine all tracks to a single GPX or CSV file.}
\usage{
combineAllTracks(
  pin,
  pout,
  trkinfo = NULL,
  type = c("both", "gpx", "csv"),
  fnm = "All Tracks"
)
}
\arguments{
\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the sanitized GPX files.}

\item{trkinfo}{A data frame that contains information about each track.}

\item{type}{A string that indicates the type of file to create ... \code{gpx},  \code{csv}, or  \code{both}.}

\item{fnm}{A filename sans extension for the resultant file (a ".gpx" or ".csv" will be added as appropriate).}
}
\value{
None, used for side effect of writing files to the  \code{pout} directory.
}
\description{
Combine all GPX track files in a directory into a single GPX or CSV file.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
