% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGPX.R
\name{combineTracks2GPX}
\alias{combineTracks2GPX}
\title{Combine tracks into a single GPX file.}
\usage{
combineTracks2GPX(pin, pout, fnm, IDs = NULL)
}
\arguments{
\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the single resultant GPX file.}

\item{fnm}{A filename sans extension for the resultant file (a ".gpx" will be added as appropriate).}

\item{IDs}{a character vector of track IDs that should be combined into a single GPX file. If \code{NULL} then all GPX files in \code{pin} will be combined or if \code{fnm} exists in \code{pout} then all GPX files with modification dates after the modification date for \code{fnm} will be appended to \code{fnm}.}
}
\value{
None, used for side effect of writing a file to the \code{pout} directory.
}
\description{
Combine given GPX track files into a single GPX file. If the single GPX file exists then tracks can be appended.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
