% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpxhelpers-utils.R
\name{compareFiles2Info}
\alias{compareFiles2Info}
\title{Compare file names in a directory to names in the information file.}
\usage{
compareFiles2Info(pin, trkinfo)
}
\arguments{
\item{pin}{Path to directory with track files.}

\item{trkinfo}{Database with track information.}
}
\value{
None, but informative messages will be written as side effects
}
\description{
Compare file names in a directory to names in the information file to see if any tracks are missing in either place.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
