% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpxhelpers-utils.R
\name{distAlongTrack}
\alias{distAlongTrack}
\title{Calculate distance along the track.}
\usage{
distAlongTrack(
  d,
  vars = c("Longitude", "Latitude"),
  units = c("miles", "km", "m")
)
}
\arguments{
\item{d}{Data frame that contains the track coordinates.}

\item{vars}{A string vector with the names in  \code{d} that contain the longitude and latitude coordinars (in that order) along the track.}

\item{units}{A string with the desired units for the distances.}
}
\value{
A numeric vector with cumulative distances along the track.
}
\description{
Calculates the cumulative distance along the track.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link[geosphere]{distGeo}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
