% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{findFileStatus}
\alias{findFileStatus}
\title{Determine track status.}
\usage{
findFileStatus(trkinfo, pin, pout)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the sanitized GPX files.}
}
\value{
Returns \code{trkinfo} with new \code{modin} for modification time of GPX in \code{pin}, \code{modout} for modification time of GPX in \code{pout}, and \code{status} for \dQuote{status} of GPX file in \code{pin} (i.e., \dQuote{NEW}, \dQuote{MODIFIED}, or \dQuote{EXISTING}).
}
\description{
Determine if the GPX files in \code{pin} are \dQuote{NEW} (i.e., never sanitized), \dQuote{MODIFIED} (i.e., GPX in \code{pin} is more recent than GPX in \code{pout}), or \dQuote{EXISTING} (i.e.,GPX is in \code{pin} and is not more recent than sanitized GPX in \code{pout}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
