% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpxhelpers-internals.R
\name{iBaseMap}
\alias{iBaseMap}
\alias{iBearing}
\alias{iBoundBaseMap}
\alias{iMakeDescription}
\alias{iOrderWalk}
\alias{iRetClr}
\alias{iSwapFromTo}
\alias{iWalkSumPts}
\title{Internal functions.}
\usage{
iBaseMap(dat)
}
\description{
Internal functions that are common to several functions in gpxhelpers.
}
\keyword{internal}
