% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCarousel.R
\name{makeCarousel}
\alias{makeCarousel}
\title{Make a carousel of images}
\usage{
makeCarousel(
  id,
  path,
  imgs = NULL,
  caps = NULL,
  rev = FALSE,
  width = c("100", "75", "50"),
  center = TRUE,
  captions = TRUE,
  controls = TRUE,
  indicators = FALSE,
  interval = 2500,
  fade = FALSE,
  hover = TRUE
)
}
\arguments{
\item{id}{Name for the carousel (make unique if more than one carousel on a page).}

\item{path}{A relative path to the folder with the images. If \code{imgs} is \code{NULL} then the folder should only contain image files that will form the carousel.}

\item{imgs}{A vector of file names with the images to form the carousel. See details.}

\item{caps}{A vector of captions for each image in the carousel. If \code{NULL} then captions will be the filename sans extension.}

\item{rev}{A logical for whether the given order of \code{imgs} and \code{caps} should be reversed. Most logically used if \code{imgs} and \code{caps} are not specified.}

\item{width}{The relative width (as a percentage) of the carousel (can only be 100, 75, or 50).}

\item{center}{Logical for whether the carousel should be centered on the page or not.}

\item{captions}{Logical for whether captions should be added.}

\item{controls}{Logical for whether the Previous and Next control items should be used.}

\item{indicators}{Logical for whether indicator buttons should be shown for the image list.}

\item{interval}{The amount of time between images on the carousel.}

\item{fade}{Logical for whether there is a slide (\code{FALSE}) or fade between images.}

\item{hover}{Logical for whether the carousel should stop on a mouseover.}
}
\value{
None, used for side effect of writing HTML code that produces a slide carousel of images.
}
\description{
Makes HTML code to be used in a quarto document that will produce a carousel of images. Names of files with images may be given in \code{imgs} (and must be in \code{path}) or if \code{imgs} is \code{NULL} then all files in \code{path} will be used to make the carousel.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
