% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{makeMasterCSV}
\alias{makeMasterCSV}
\title{Write all track information into a master CSV file.}
\usage{
makeMasterCSV(trkinfo, pin, pout, fnm, msgcutoff = 100)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the single resultant GPX file.}

\item{fnm}{A filename sans extension that contains the name for the output CSV file.}

\item{msgcutoff}{A numeric that indicates the maximum number of tracks to be sanitized before an overall progress bar will be given (rather than individual notices).}
}
\value{
Will invisibly return the data frame that is also written to the CSV file in the same directory.
}
\description{
Combine track information with track GPX data into a single CSV file.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
