% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{makeMasterGPX}
\alias{makeMasterGPX}
\title{Combine tracks into a single master GPX file.}
\usage{
makeMasterGPX(trkinfo, pin, pout, fnm, msgcutoff = 100)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the single resultant GPX file.}

\item{fnm}{A filename sans extension for the resultant file (a ".gpx" will be added as appropriate).}

\item{msgcutoff}{A numeric that indicates the maximum number of tracks to be sanitized before an overall progress bar will be given (rather than individual notices).}
}
\value{
None, used for side effect of writing a file to the \code{pout} directory.
}
\description{
Combine given GPX track files into a single master GPX file. If the single GPX file exists then tracks will be appended.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
