% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageHelpers.R
\name{makeThumbs}
\alias{makeThumbs}
\title{Make thumbnails of images}
\usage{
makeThumbs(id, path, ncol = 6, imgs = NULL, caps = NULL, rev = FALSE)
}
\arguments{
\item{id}{Name for the thumbnail gallery.}

\item{path}{A relative path to the folder with the images. If \code{imgs} is \code{NULL} then the folder should only contain image files that will form the carousel.}

\item{ncol}{A numeric that indicates the number of columns to be used for the thumbnails. Defaults to 6. Use a bigger number to make smaller thumbnails.}

\item{imgs}{A vector of file names with the images to form the thumbnails. See details.}

\item{caps}{A vector of captions for each image in the thumbnails. If \code{NULL} then captions will be the filename sans extension.}

\item{rev}{A logical for whether the given order of \code{imgs} and \code{caps} should be reversed. Most logically used if \code{imgs} and \code{caps} are not specified.}
}
\value{
None, used for side effect of writing HTML code that produces thumbnails of images.
}
\description{
Makes HTML code to be used in a quarto document that will produce a thumbnail of images that can then be clicked on to view as a lightbox. Names of files with images may be given in \code{imgs} (and must be in \code{path}) or if \code{imgs} is \code{NULL} then all files in \code{path} will be used to make the carousel.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
