% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walks.R
\name{makeWalkGPX}
\alias{makeWalkGPX}
\title{Combine tracks for a walk into a single GPX file.}
\usage{
makeWalkGPX(pin, pout, fnm, IDs2Add)
}
\arguments{
\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the single resultant GPX file.}

\item{fnm}{A filename sans extension for the resultant file (a ".gpx" will be added as appropriate).}

\item{IDs2Add}{A character vector of track IDs for a walk that should be combined into a single GPX file.}
}
\value{
None, used for side effect of writing a file to the \code{pout} directory.
}
\description{
Combine given GPX track files into a single GPX file.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
