% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{renameGPXfile}
\alias{renameGPXfile}
\title{Rename tracks.}
\usage{
renameGPXfile(fold, fnew, pth, deleteOld = FALSE)
}
\arguments{
\item{fold}{Name for \dQuote{old} GPX file.}

\item{fnew}{Name for \dQuote{new} GPX file.}

\item{pth}{Path after the working directory that contains the \dQuote{old} GPX files and will receive the \dQuote{new} GPX files.}

\item{deleteOld}{A logical that indicates whether \code{fold} should be deleted from \code{pth} after \code{fnew} is created. Defauts to \code{FALSE} (i.e., not deleting \code{fold}).}
}
\value{
None, used for side effect of writing new GPX files to the \code{pth} directory (and possibly deleting now old GPX files from the same directory).
}
\description{
Rename gpx filenames but also, importantly, adjusting the track name in the \code{<name>} tag.
}
\details{
NONE YET
}
\examples{
\dontrun{
## Set working directory to where the files are
## Single file
renameGPXfile("NCTBF01","NCTBF001")

## Example of multiple files with a consistent file change pattern
olds <- c("NCTBF01","NCTBF07a")
news <- stringr::str_replace(olds,"NCTBF","NCTBF0")
for (i in seq_along(olds)) renameGPXfile(olds[i],news[i],deleteOld=TRUE)
}

}
\author{
Derek H. Ogle
}
\keyword{manip}
