% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walks.R
\name{walkElevation}
\alias{walkElevation}
\title{Visualize elevations for tracks that form a walk.}
\usage{
walkElevation(walkdat, title = NULL, elev_bufr = 0.01, maval = 10)
}
\arguments{
\item{walkdat}{A data frame that contains tracks that form a contiguous \dQuote{walk}. Usually made with \code{\link{walkMaker}}.}

\item{title}{A string for the map title.}

\item{elev_bufr}{A numeric that makes a slight buffer around the elevation data.}

\item{maval}{The number of points to include in the moving average. Use 1 to maintain raw data. Defaults to 10.}
}
\value{
Returns a \code{ggplot2} object. The object will be displayed if the returned object is not assigned to name.
}
\description{
Plot elevations for tracks in a data frame that form a contiguous \dQuote{walk}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link{walkMap}} \code{\link{walkSummary}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
