% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkGetTrackIDs.R
\name{walkGetTrackIDs}
\alias{walkGetTrackIDs}
\alias{walkGetPhotos}
\title{Extract track IDs and photos for a walk.}
\usage{
walkGetTrackIDs(d, whichWalk)

walkGetPhotos(d, walkIDs, path = "Images")
}
\arguments{
\item{d}{A data frame from either the \dQuote{Walks} or \dQuote{Photos} tab in the trail mapping info sheet.}

\item{whichWalk}{A numeric that identifies the code for the \dQuote{walk}.}

\item{walkIDs}{A vector of track name strings that represent a contiguous \dQuote{walk}. This may be returned from \code{\link{walkGetTrackIDs}}.}

\item{path}{A directory from the working directory that contains the photo image files.}
}
\value{
\code{walkGetTrackIDs} returns a vector of track IDs for the walk and \code{walkGetPhotos} returns a data frame that contains the filename (with the \code{path} prepended) and caption for photos from the walk.
}
\description{
Extract track IDs and photos for a \dQuote{walk}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link{walkGetTrackIDs}}, \code{\link{walkMap}}, and \code{\link{walkSummary}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
