% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkMaker.R
\name{walkMaker}
\alias{walkMaker}
\title{Concatenate tracks into a walk.}
\usage{
walkMaker(
  trkdata,
  trkinfo,
  walkIDs,
  startIDs = walkIDs[1:2],
  findOrder = FALSE,
  basedate = "2022-01-01",
  verbose = FALSE
)
}
\arguments{
\item{trkdata}{A data frame that contains coordinates and information about each track.}

\item{trkinfo}{A data frame that contains information about each track.}

\item{walkIDs}{A vector of track name strings that can be appended into a contiguous \dQuote{walk}. Generally first two names should be the order of the first two tracks in the \dQuote{walk}.}

\item{startIDs}{A length two vector of track name strings that indicate the order of the first two tracks in the \dQuote{walk}. Defaults to the first two items in \code{walkIDs}.}

\item{findOrder}{A logical that indicates that the track names in \code{walkIDs} are NOT in walking order and an attempt should be made to place the tracks in a walking order based on the \code{connected} field in \code{trkinfo}. See notes below.}

\item{basedate}{A string with a date that will serve as the base date for the dummy times in the returned data frame. Defaults to "2022-01-01".}

\item{verbose}{A logical for whether the progress of connecting tracks should be displayed or not.}
}
\value{
A date frame with the same variables as \code{trkdata}, but only with the tracks in \code{walkIDs} and in order of the \dQuote{walk}.
}
\description{
Concatenate a vector of track names that represent a contiguous \dQuote{walk} into a data frame.
}
\details{
NONE YET
}
\note{
If any of the track names in \code{walkIDs} are repeated because the \dQuote{walk} will include repeated walkings of some segments then the user must put the track names in the order to be walked in \code{walkIDs} and \code{findOrder} must be \code{FALSE}.
}
\examples{
## None yet.
}
\seealso{
\code{\link{walkMap}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
