% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkReports.R
\name{walkReports}
\alias{walkReports}
\title{Write walk report HTML files.}
\usage{
walkReports(
  walk,
  project,
  datfile,
  basedir,
  tmplt = "Walk_Template.Rmd",
  OMap_type = "bing"
)
}
\arguments{
\item{walk}{A vector of walk codes.}

\item{project}{A project name.}

\item{datfile}{The name of the file in \code{pth} that contains all of the track information.}

\item{basedir}{A path string to where \code{datfile}, the folder with the images, and the folder in which to put the resultant HTML file reside.}

\item{tmplt}{A name for the template to use.}

\item{OMap_type}{Type of OpenStreetMap to use.}
}
\value{
None, but an html file will be created.
}
\description{
Write \dQuote{walk} report HTML files.
}
\details{
NONE YET
}
\examples{
\dontrun{
project <- "Bayfield County"
basedir <- file.path("C:/aaaPersonal/Maps_GPS",project)
datfile <- paste0(project,".csv")
walk <- "RYND1"
walkReports(walk,project,datfile,basedir)
}

}
\author{
Derek H. Ogle
}
\keyword{manip}
