% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkSummary.R
\name{walkSummary}
\alias{walkSummary}
\title{Summarize tracks that form a walk.}
\usage{
walkSummary(walkdata)
}
\arguments{
\item{walkdata}{A data frame that contains tracks that form a contiguous \dQuote{walk}. Usually made with \code{\link{walkMaker}}.}
}
\value{
Returns a \code{kable} table object.
}
\description{
Create a table that summarizes tracks in a data frame that form a contiguous \dQuote{walk}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link{walkMap}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
