% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGPX.R
\name{writeGPXnInfo2CSV}
\alias{writeGPXnInfo2CSV}
\title{Write track information into a CSV file.}
\usage{
writeGPXnInfo2CSV(trkinfo, fnm)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{fnm}{A filename (with \dQuote{.gpx} extension) that contains the tracks to output to a CSV file with same base name.}
}
\value{
Will invisibly return the data frame that is also written to the CSV file in the same directory.
}
\description{
Combine track information with track GPX data into a single CSV file.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
