% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpxhelpers-internals.R
\name{iWalkSumPts}
\alias{iWalkSumPts}
\title{Internal functions.}
\usage{
iWalkSumPts(walkdat)
}
\description{
Internal functions that are common to several functions in gpxhelpers.
}
\keyword{internal}
