% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileUtils.R
\name{renameGPXfile}
\alias{renameGPXfile}
\title{Rename tracks.}
\usage{
renameGPXfile(fold, fnew, pnew = "renamed")
}
\arguments{
\item{fold}{Name for \dQuote{old} GPX file.}

\item{fnew}{Name for \dQuote{new} GPX file.}

\item{pnew}{Path after the working directory to contain the \dQuote{new} GPX files.}
}
\value{
None, used for side effect of writing new GPX files to the \code{pth} directory (and possibly deleting now old GPX files from the same directory).
}
\description{
Rename gpx filenames but also, importantly, adjusting the track name in the \code{<name>} tag.
}
\details{
NONE YET
}
\examples{
\dontrun{
## Set working directory to where the files are
## Single file
renameGPXfile("NCTBF01","NCTBF001")

## Example of multiple files with a consistent file change pattern
olds <- c("NCTBF01","NCTBF07a")
news <- stringr::str_replace(olds,"NCTBF","NCTBF0")
cbind(olds,news)
renameGPXfile(olds,news)

## Example of multiple files to get new numbers padded with zeroes
olds <- c("NCTBF01","NCTBF07a")
news <- paste0("NCTBF",stringr::str_pad(seq_along(olds),width=3,pad=0))
cbind(olds,news)
renameGPXfile(olds,news)
}

}
\author{
Derek H. Ogle
}
\keyword{manip}
