% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeTracks.R
\name{sanitizeTracks}
\alias{sanitizeTracks}
\title{Sanitize tracks.}
\usage{
sanitizeTracks(trkinfo, pin, pout, basedate = NULL)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the sanitized GPX files.}

\item{basedate}{A string with a date that will serve as the base date for the dummy times in the sanitized GPX file. Defaults to "2022-01-01".}
}
\value{
None, used for side effect of writing sanitized GPX tracks to the \code{pout} directory.
}
\description{
Sanitize all GPX track files in a directory by removing the \code{type} and  \code{extension} fields, adding a \code{desc}ription field, and replacing \code{time} with dummy times.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
