% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walkMap.R
\name{walkMap}
\alias{walkMap}
\alias{walkElevation}
\title{Visualize tracks that form a walk.}
\usage{
walkMap(
  walkdata,
  title = NULL,
  label_tracks = TRUE,
  OMap_type = c("esri-topo", "esri", "apple-iphoto", "bing", "stamen-terrain", "none"),
  OMap_bufr = 2e-05
)

walkElevation(walkdata, title = NULL, elev_bufr = 0.01)
}
\arguments{
\item{walkdata}{A data frame that contains tracks that form a contiguous \dQuote{walk}. Usually made with \code{\link{walkMaker}}.}

\item{title}{A string for the map title.}

\item{label_tracks}{A logical for whether (or not) the tracks should be labeled with a unique number.}

\item{OMap_type}{A string for the type of OpenStreet map to use under the track paths. Use \code{"none"} to not using an OpenStreet map.}

\item{OMap_bufr}{A numeric that makes the OpenStreet map slightly larger than the space that the track paths require.}

\item{elev_bufr}{A numeric that makes a slight buffer around the elevation data.}
}
\value{
Returns a \code{ggplot2} object that is the map. The map will be displayed if the returned object is not assigned to name.
}
\description{
Create a map or plot elevations for tracks in a data frame that form a contiguous \dQuote{walk}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link{walkSummary}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
