% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{walkMap}
\alias{walkMap}
\alias{allTracksMap}
\title{Visualize tracks that form a walk.}
\usage{
walkMap(
  dat,
  clrs,
  LAT_bottom = NULL,
  LAT_top = NULL,
  LON_left = NULL,
  LON_right = NULL,
  map_bufr = 4e-05,
  label_tracks = TRUE,
  verbose = TRUE
)

allTracksMap(
  dat,
  clrs,
  LAT_bottom = NULL,
  LAT_top = NULL,
  LON_left = NULL,
  LON_right = NULL,
  map_bufr = 4e-05,
  walk = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{A data frame that contains tracks to map, contiguous tracks for \code{walkMap} that are usually made with \code{\link{walkMaker}} and all tracks for \code{allTracksMap}.}

\item{clrs}{A named vector of colors that will be applied to identify tracks by \code{Type} in \code{dat}.}

\item{LAT_bottom}{A latitude coordinate for the bottom of the bounding box for the map. Defaults to just below the minimum latitude found in \code{trkdata}.}

\item{LAT_top}{A latitude coordinate for the top of the bounding box for the map. Defaults to just above the maximum latitude found in \code{trkdata}.}

\item{LON_left}{A longitude coordinate for the left-side of the bounding box for the map. Defaults to just left of the minimum latitude found in \code{trkdata}.}

\item{LON_right}{A longitude coordinate for the right-side of the bounding box for the map. Defaults to just right the maximum latitude found in \code{trkdata}.}

\item{map_bufr}{A numeric that makes the map slightly larger than the space that the track paths require.}

\item{label_tracks}{A logical for whether (or not) the tracks should be labeled with a unique number in \code{walkMap} or the track ID in \code{allTracksMap}.}

\item{verbose}{A logical for whether progress messages should be displayed or not.}

\item{walk}{A ggplot2 object made with \code{walkMap} that will be used to highlight the \dQuote{walk} on the map of all tracks.}
}
\value{
Returns a \code{leaflet} object that is the map. The map will be displayed if the returned object is not assigned a name.
}
\description{
Create a map or plot elevations for tracks in a data frame that form a contiguous \dQuote{walk}.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\seealso{
\code{\link{walkSummary}}
}
\author{
Derek H. Ogle
}
\keyword{manip}
