% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walks.R
\name{walkReports}
\alias{walkReports}
\title{Write walk report HTML files.}
\usage{
walkReports(
  walk,
  project,
  datfile,
  basedir,
  tmplt = "Walk_Template.Rmd",
  showFileInBrowser = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{walk}{A vector of walk codes.}

\item{project}{A project name.}

\item{datfile}{The name of the file in \code{pth} that contains all of the track information.}

\item{basedir}{A path string to where \code{datfile}, the folder with the images, and the folder in which to put the resultant HTML file reside.}

\item{tmplt}{A name for the template to use.}

\item{showFileInBrowser}{A logical for whether to open the resultant file in the broswer or not (default is to not).}

\item{quiet}{A logical for whether the progress of processing the markdown file should be shown (default is to not).}
}
\value{
None, but an html file will be created.
}
\description{
Write \dQuote{walk} report HTML files.
}
\details{
NONE YET
}
\examples{
\dontrun{
project <- "Bayfield County"
basedir <- file.path("C:/aaaPersonal/MAPPING",project)
datfile <- paste0(project,".csv")
walk <- "FR4138051"
walkReports(walk,project,datfile,basedir,showFileInBrowser=TRUE)
}

}
\author{
Derek H. Ogle
}
\keyword{manip}
