% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineGPX.R
\name{writeGPXnInfo2CSV}
\alias{writeGPXnInfo2CSV}
\title{Write track information into a CSV file.}
\usage{
writeGPXnInfo2CSV(trkinfo, pin, pout, fnm, IDs = NULL)
}
\arguments{
\item{trkinfo}{A data frame that contains information about each track.}

\item{pin}{Path after the working directory that contains the original GPX files.}

\item{pout}{Path after the working directory to put the single resultant GPX file.}

\item{fnm}{A filename sans extension that contains the name for the output CSV file.}

\item{IDs}{a character vector of track IDs that should be combined into a single GPX file. If \code{NULL} then all GPX files in \code{pin} will be combined or if \code{fnm} exists in \code{pout} then all GPX files with modification dates after the modification date for \code{fnm} will be appended to \code{fnm}.}
}
\value{
Will invisibly return the data frame that is also written to the CSV file in the same directory.
}
\description{
Combine track information with track GPX data into a single CSV file.
}
\details{
NONE YET
}
\examples{
## None yet.

}
\author{
Derek H. Ogle
}
\keyword{manip}
